Revert "Merge branch 'eventstore-phase-3' into cleanup-setup"
This reverts commit 77fddd4d285c767964f91ed58ec0a64f6a7f445c, reversing
changes made to b2661ab351895d167a35be7c16c6e14377cd042c.

+++ internal/query/projection/config.go
@@ -7,7 +7,7 @@ import (
 type Config struct {
 	RequeueEvery          time.Duration
 	RetryFailedAfter      time.Duration
-	MaxFailureCount       uint8
+	MaxFailureCount       uint
 	ConcurrentInstances   uint
 	BulkLimit             uint64
 	Customizations        map[string]CustomConfig
